Refactoring upgrade_analyzer
